Move date/time types to schema file
